client: use 64-bit value in sync read eof logic
The file size can jump to a value that is very much larger than our current
position (for example, it could be a disk image file that gets a sparse
write at a large offset).  Use a 64-bit value so that 'some' doesn't
overflow.

+++ src/client/Client.cc
@@ -5865,7 +5865,7 @@ int Client::_read_sync(Fh *f, uint64_t off, uint64_t len, bufferlist *bl)
     if (r >= 0 && r < wanted) {
       if (pos < in->size) {
 	// zero up to known EOF
-	int some = in->size - pos;
+	uint64_t some = in->size - pos;
 	if (some > left)
 	  some = left;
 	bufferptr z(some);
